Bug fixes and minor updates for control bar
